Move driTrait to traits dir
